refactor: extract workflow FSM expansion
