+++ tests/workflow-fsm-expansion.test.ts
@@ -6,11 +6,18 @@ import { afterEach, describe, expect, it } from "vitest";
 import { BUILTIN_WORKFLOW_TEMPLATES } from "../src/builtin-templates.js";
 import { decideNextStep } from "../src/lifecycle/state-machine-dispatch.js";
 import {
+  explainWorkflow as explainWorkflowFromFacade,
+  loadExpandedWorkflow as loadExpandedWorkflowFromFacade,
+  loadProjectWorkflow as loadProjectWorkflowFromFacade,
+  validateExpandedWorkflowReferences as validateExpandedWorkflowReferencesFromFacade
+} from "../src/workflow.js";
+import {
   explainWorkflow,
   loadExpandedWorkflow,
+  loadProjectWorkflow,
   validateExpandedWorkflowReferences
-} from "../src/workflow.js";
-import type { ExpandedWorkflow } from "../src/workflow.js";
+} from "../src/workflow/fsm-expansion.js";
+import type { ExpandedWorkflow } from "../src/workflow/types.js";
 
 const tempRoots: string[] = [];
 
@@ -26,6 +33,17 @@ afterEach(async () => {
       rm(root, { force: true, recursive: true })
     )
   );
+});
+
+describe("workflow FSM expansion facade", () => {
+  it("keeps FSM expansion exports available through the workflow facade", () => {
+    expect(loadExpandedWorkflowFromFacade).toBe(loadExpandedWorkflow);
+    expect(loadProjectWorkflowFromFacade).toBe(loadProjectWorkflow);
+    expect(validateExpandedWorkflowReferencesFromFacade).toBe(
+      validateExpandedWorkflowReferences
+    );
+    expect(explainWorkflowFromFacade).toBe(explainWorkflow);
+  });
 });
 
 describe("state machine workflow definitions", () => {
